Use the latest OkHttp public APIs
There is no need to manually set timeouts on the internally generated
OkHttpClient instances because they now have default timeouts.

The MockWebServer can now be used as a JUnit rule avoiding the need to
start and stop instances manually.

We should depend on the new HttpUrl implementation.

Also, avoid setting the connection pool in
executorServiceIsIntrumented().