bugfix: Rewrite network request instrumentation w/ interceptors.
`OkHttpClient` relies on a `Dispatcher` for executing asynchronous
requests.  `Dispatcher` executes these async requests on an
`ExecutorService`.  Based on this fact, we were able to provide an
instrumented `ExecutorService` to the dispatcher in order to record
metrics on any requests.  The issue is that the instrumented executor
service is only used when executing async requests - it is not used
when executing sync requests (via `Call.execute()`).

This change replaces the instrumented executor service with an
instrumented network interceptor.  By using interceptors, we can record
metrics on request regardless of whether they are synchronous or
asynchronous.

Closes #43